solves bug with init
init was not storing logfiles in the correct folder by default. This isn't an issue with the code, but with a bug in STATA in Windows, I suspect. The previous code behaved as expected in Linux but failed to find the log folder in Windows. Solved by using Mata code.